Simple Objectpooler / Restructure Project
Firstfully I decided to Focus  on Creating Essential Modules for future implementation and generally to create my own library to use for incoming projects.
This commit adds the ObjectPooler class to the project. The ObjectPooler class is a script that allows for the pooling of game objects, in order to reduce the number of instantiations and destructions that occur during gameplay. This can help to improve performance and reduce memory usage.

The ObjectPooler class contains a pools array, which stores all of the different object pools that will be used during gameplay. Each object pool is defined by a Pool class, which contains a poolName string and a prefab game object.

Additionally, the ObjectPooler class contains a poolDictionary dictionary, which maps a poolName string to the corresponding index of the pools array. This is used to quickly access a specific object pool when needed.

The ObjectPooler class also has a TestFunction method that can be used to test the pooling functionality. This method is called when a specific event occurs.

The Initialize method is used to pre-instantiate a certain amount of game objects from each pool and add the pool name and index to the poolDictionary. This is done in order to reduce the amount of lag that might occur during gameplay.

The SelectPool method is used to retrieve a specific object pool based on the poolName parameter passed to it. This method returns the Pool object that corresponds to the specified poolName.

The InitObject method is used to initialize a new game object for the specified Pool. This method instantiates the prefab and sets its active state based on the state parameter passed to it. The new game object is added to the Pool object's objectPool list.

Finally, the PoolObject method is used to retrieve a game object from the specified Pool. If there are no inactive game objects available, a new game object is initialized using the InitObject method. The PoolObject method returns the retrieved or newly created game object.

This commit also includes the addition of a static Instance variable that is used to store a reference to the ObjectPooler object in order to allow for easy access from other scripts.
In addition to the above methods, the ObjectPooler class also includes OnEnable and OnApplicationQuit methods that are used to subscribe and unsubscribe from an event defined in the InputManager script.

The ObjectPooler script can be attached to a game object in a scene, and can then be used to pool and retrieve game objects during gameplay. The Pool class can be extended to include additional variables and functionality as needed for specific use cases.

Overall, this commit lays the foundation for implementing object pooling in the project, which can help to improve performance and reduce memory usage.