Simple AudioManager
Add AudioManager_Test class

This commit adds the `AudioManager_Test` class, which is responsible for managing various audio sources in the game. The class has functions to play and stop background music, ambient sound, and a looping track. It also has a function to play a one-shot sound effect.

The class is a MonoBehaviour and is designed to be attached to a game object in the scene. It uses Unity's built-in `AudioSource` component to play audio.

Specifically, the `AudioManager_Test` class has the following features:
- A `loopClip` AudioClip that can be played in a loop using the `PlayLoop` function. The loop can be toggled on/off using the `ToggleLoop` function.
- A `musicClip` AudioClip that can be played in a loop using the `PlayMusic` function. The music can be toggled on/off using the `ToggleMusic` function.
- An `ambientClip` AudioClip that can be played in a loop using the `PlayAmbient` function. The ambient sound can be toggled on/off using the `ToggleAmbient` function.
- A `oneShotClip` AudioClip that can be played once using the `PlayOneShot` function.

All audio sources are initialized in the `Awake` function and stored as private variables for easy access.

Note: This is an initial commit and the class is still a work in progress. Future commits will refine the implementation and add more features.